Fix usage of latex free last names
