+++ src/main/java/org/jabref/model/groups/LastNameGroup.java
@@ -1,7 +1,13 @@
 package org.jabref.model.groups;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.jabref.model.entry.Author;
+import org.jabref.model.entry.AuthorList;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.Field;
 
@@ -13,10 +19,21 @@ public class LastNameGroup extends KeywordGroup {
         super(name, context, searchField, searchExpression, true);
     }
 
+    static List<String> getAsLastNamesLatexFree(Field field, BibEntry bibEntry) {
+        return bibEntry.getField(field)
+                       .map(AuthorList::parse)
+                       .map(AuthorList::getAuthors)
+                       .map(authors ->
+                               authors.stream()
+                                      .map(Author::getLastLatexFree)
+                                      .flatMap(Optional::stream)
+                                      .collect(Collectors.toList()))
+                       .orElse(Collections.emptyList());
+    }
+
     @Override
     public boolean contains(BibEntry entry) {
-        return AutomaticPersonsGroup.getAsLastNamesLatexFree(searchField, entry)
-                                    .stream().anyMatch(name -> name.equals(searchExpression));
+        return getAsLastNamesLatexFree(searchField, entry).stream().anyMatch(name -> name.equals(searchExpression));
     }
 
     @Override
